+++ src/Manager.php
@@ -7,8 +7,7 @@ use Cyberaxio\GitManager\Commands\GitCommand;
 class Manager
 {
 	private function __construct()
-	{		
-	}
+	{}
 
 	public static function new($config = [])
 	{
@@ -22,10 +21,27 @@ class Manager
 
 	public static function init($path = null)
 	{
+		if(!realpath($path)){
+			mkdir($path, 0777, true);
+		}
 		(new GitCommand('git init'))
 		->setCwd(realpath($path))
 		->run();
 
 		return self::find($path);
+	}
+
+	public static function clone($url, $path, $privateKey = null, $port = 22)
+	{
+		$repository = self::find($path, ['cloning' => true]);
+		if ($repository->exists()) {
+			return $repository;
+		}
+		$repository->setUrl($url);
+		$repository->setPort($port);
+		$repository->setPrivateKey($privateKey);
+		$repository->clone($path);
+
+		return $repository;
 	}
 }